Use AsideTableOfContent to ensure same component and padding + cleanup and code enhacements
