+++ src/pages/about.js
@@ -8,7 +8,7 @@ import { getSitePage } from '../cms';
 import AboutContent from '../components/AboutContent';
 import ContentPage from '../components/ContentPage';
 import Page from '../components/Page';
-import TableOfContent from '../components/AboutContent/TableOfContent';
+import AsideTableOfContent from '../components/AsideTableOfContent';
 
 const useStyles = makeStyles({
   root: {
@@ -102,7 +102,11 @@ function About(takwimu) {
     <Page takwimu={takwimu} title={title}>
       <ContentPage
         aside={
-          <TableOfContent current={current} contentHeadings={contentHeadings} />
+          <AsideTableOfContent
+            current={current}
+            contentHeadings={contentHeadings}
+            generateHref={({ link }) => `/${link}`}
+          />
         }
         classes={{ root: classes.root }}
       >
